instruments/survey-app/requirements.txt
flask
requests
python-dotenv
flask-mail
pandas
boto3
plotly
mathplotlib